seperated some fun and var from files
