+++ lib/quickScan.dart
@@ -13,7 +13,6 @@ class QuickScanPage extends StatefulWidget {
 }
 
 class _QuickScanPageState extends State<QuickScanPage> {
-
   final _scrollController = ScrollController();
 
   String debugOutput = "";
@@ -26,23 +25,6 @@ class _QuickScanPageState extends State<QuickScanPage> {
   @override
   void initState() {
     super.initState();
-  }
-
-  void showLoadingIndicator(String text) {
-    showDialog(
-      context: context,
-      barrierDismissible: false,
-      builder: (BuildContext context) {
-        return WillPopScope(
-            onWillPop: () async => false,
-            child: AlertDialog(
-              shape: RoundedRectangleBorder(
-                  borderRadius: BorderRadius.all(Radius.circular(8.0))),
-              backgroundColor: Colors.black87,
-              content: LoadingIndicator(text: text),
-            ));
-      },
-    );
   }
 
   Widget _buildDeviceList() {
@@ -146,47 +128,47 @@ class _QuickScanPageState extends State<QuickScanPage> {
   @override
   Widget build(BuildContext context) {
     return Scaffold(
-        backgroundColor: hPi4Global.appBackgroundColor,
-        appBar: AppBar(
-          backgroundColor: hPi4Global.hpi4Color,
-          leading:
-              Consumer<BleScannerState>(builder: (context, bleScanner, child) {
-            return new IconButton(
-                icon: new Icon(Icons.arrow_back, color: Colors.white),
-                onPressed: () async {
-                  await Provider.of<BleScanner>(context, listen: false)
-                      .stopScan();
-                  Navigator.of(context).pop();
-                });
-          }),
-          title: Row(
-            mainAxisAlignment: MainAxisAlignment.end,
-            mainAxisSize: MainAxisSize.max,
-            children: <Widget>[
-              Image.asset('assets/proto-online-white.png',
-                  fit: BoxFit.fitWidth, height: 30),
-            ],
-          ),
+      backgroundColor: hPi4Global.appBackgroundColor,
+      appBar: AppBar(
+        backgroundColor: hPi4Global.hpi4Color,
+        leading:
+            Consumer<BleScannerState>(builder: (context, bleScanner, child) {
+          return new IconButton(
+              icon: new Icon(Icons.arrow_back, color: Colors.white),
+              onPressed: () async {
+                await Provider.of<BleScanner>(context, listen: false)
+                    .stopScan();
+                Navigator.of(context).pop();
+              });
+        }),
+        title: Row(
+          mainAxisAlignment: MainAxisAlignment.end,
+          mainAxisSize: MainAxisSize.max,
+          children: <Widget>[
+            Image.asset('assets/proto-online-white.png',
+                fit: BoxFit.fitWidth, height: 30),
+          ],
         ),
-        body:Scrollbar(
+      ),
+      body: Scrollbar(
+        controller: _scrollController,
+        thumbVisibility: true,
+        trackVisibility: true,
+        thickness: 10,
+        child: SingleChildScrollView(
           controller: _scrollController,
-          thumbVisibility: true,
-          trackVisibility: true,
-          thickness: 10,
-          child: SingleChildScrollView(
-            controller: _scrollController,
-            scrollDirection: Axis.vertical,
-            child: Padding(
-              padding: const EdgeInsets.all(24.0),
-              child: Column(mainAxisAlignment: MainAxisAlignment.spaceEvenly,
-                  //mainAxisSize: MainAxisSize.max,
-                  children: <Widget>[
-                    _buildDeviceList(),
-
-                  ]),
-            ),
+          scrollDirection: Axis.vertical,
+          child: Padding(
+            padding: const EdgeInsets.all(24.0),
+            child: Column(
+                mainAxisAlignment: MainAxisAlignment.spaceEvenly,
+                //mainAxisSize: MainAxisSize.max,
+                children: <Widget>[
+                  _buildDeviceList(),
+                ]),
           ),
         ),
+      ),
     );
   }
 }
