+++ lib/sizeConfig.dart
@@ -9,9 +9,9 @@ class SizeConfig {
 
   void init(BuildContext context) {
     _mediaQueryData = MediaQuery.of(context);
-     screenWidth = _mediaQueryData.size.width;
-     screenHeight = _mediaQueryData.size.height;
-     blockSizeHorizontal = screenWidth / 100;
-     blockSizeVertical = screenHeight / 100;
+    screenWidth = _mediaQueryData.size.width;
+    screenHeight = _mediaQueryData.size.height;
+    blockSizeHorizontal = screenWidth / 100;
+    blockSizeVertical = screenHeight / 100;
   }
 }
