+++ lib/utils/charts.dart
@@ -1,0 +1,52 @@
+import 'package:flutter/material.dart';
+import 'package:fl_chart/fl_chart.dart';
+import '../sizeConfig.dart';
+
+class buildPlots {
+  LineChartBarData currentLine(List<FlSpot> points, Color plotcolor) {
+    return LineChartBarData(
+      spots: points,
+      dotData: FlDotData(
+        show: false,
+      ),
+      gradient: LinearGradient(
+        colors: [plotcolor, plotcolor],
+        //stops: const [0.1, 1.0],
+      ),
+      barWidth: 3,
+      isCurved: false,
+    );
+  }
+
+  buildChart(int vertical, int horizontal, List<FlSpot> source, Color plotColor) {
+    return Container(
+      height: SizeConfig.blockSizeVertical * vertical,
+      width: SizeConfig.blockSizeHorizontal * horizontal,
+      child: LineChart(
+        LineChartData(
+          lineTouchData: LineTouchData(enabled: false),
+          clipData: FlClipData.all(),
+          gridData: FlGridData(
+            show: true,
+            drawVerticalLine: false,
+            drawHorizontalLine: false,
+          ),
+          borderData: FlBorderData(
+            show: false,
+            //border: Border.all(color: const Color(0xff37434d)),
+          ),
+          titlesData: FlTitlesData(
+            leftTitles: AxisTitles(sideTitles: SideTitles(showTitles: false)),
+            rightTitles: AxisTitles(sideTitles: SideTitles(showTitles: false)),
+            bottomTitles: AxisTitles(sideTitles: SideTitles(showTitles: false)),
+            topTitles: AxisTitles(sideTitles: SideTitles(showTitles: false)),
+          ),
+          lineBarsData: [
+            currentLine(source, plotColor),
+          ],
+        ),
+        swapAnimationDuration: Duration.zero,
+      ),
+    );
+  }
+}
